architecture changement, only one global dag now
